// src/app/pages/cabinet/moders/enums/permission.enum.ts
export enum PermissionEnum {
  studentsGet = 'StudentsGet',
  studentsSave = 'StudentsSave',
  studentsDelete = 'StudentsDelete',

  facesGet = 'FacesGet',
  facesSave = 'FacesSave',
  facesDelete = 'FacesDelete',

  assistancesGet = 'AssistancesGet',
  assistancesSave = 'AssistancesSave',
  assistancesDelete = 'AssistancesDelete',

  newsGet = 'NewsGet',
  newsSave = 'NewsSave',
  newsDelete = 'NewsDelete',
}